pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "krisp-ai-engineer-role-task"
version = "0.1.0"
description = "LLM-based question clustering pipeline (dedupe → cluster → classify)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "faiss-cpu>=1.7.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
    "scikit-learn>=1.3.0",
]

[project.scripts]
run-pipeline = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
